orka: Speed up loading .gltf files
A JSON.Streams.Stream object that uses a file stream will perform a
'Read operation for each character that must be read. Instead, by
determining the file size, we can use a single 'Read operation to read
the whole file in one go and store its content in a String in memory.

Signed-off-by: onox <denkpadje@gmail.com>