orka: Position the camera at the origin to reduce floating point errors
Putting large numbers (translation of the positions) in both the view
and world matrices will reduce the precision during floating point
operations.

These big translations can be removed from the matrices by positioning
the camera at (or near) the origin and computing the difference between
the camera and the model. The root node of the model is then translated
by this difference.

Signed-off-by: onox <denkpadje@gmail.com>