orka: Add job graph processing system
The package Orka.Jobs.Workers contains a number of Worker tasks which
try to dequeue and process Jobs from a shared Queue.

Each job A can have zero or one dependent job B (thus A is a dependency
of B). Such jobs have priority over jobs that are not a dependency of
some other job.

After a job A has been executed by a worker, the worker will decrement
the number of dependencies in dependent job B (if there is one). If B's
number of dependencies gets reduced to zero, then B will be enqueued to
the shared queue so that it can be processed later by a worker.

Signed-off-by: onox <denkpadje@gmail.com>