orka: Add Parallel_Job interface for parallelizable jobs
A tagged type can inherit Slice_Job (Which implements the Parallel_Job
interface) and be made parallelizable with the Parallelize function.

This function will return a pointer to a job which (when it is executed
by a worker) will spawn multiple jobs of the (given) Parallel_Job'Class
(each with a different range) and enqueue them. These jobs will
effectively run in parallel if there are multiple workers.

For example:

   type Some_Parallel_Job is new Orka.Jobs.Slice_Job with null record;

   overriding
   procedure Execute (Object : Some_Parallel_Job; From, To : Positive);

An instance of this job can then be created and parallelized as follows:

   Job_1 : Orka.Jobs.Job_Ptr := new Some_Jobs.Some_Parallel_Job;

   Job_2 : Orka.Jobs.Job_Ptr := Orka.Jobs.Parallelize
     (Orka.Jobs.Parallel_Job'Class (Job_1.all)'Access, 24, 6);

When Job_2 is enqueued (Job_1 must not be enqueued!) it will spawn 4
jobs with the ranges 1..6, 7..12, 13..18, and 19..24.

Job_2 can be used with/in Set_Dependencies.

Signed-off-by: onox <denkpadje@gmail.com>