orka: Add package Resources.Loader to asynchronously load resources
The function Load returns a constant reference (accessor) which (when
implicitly dereferenced) is a pointer to a Future'Class. This is a
protected object that can be used to wait until the resource has been
loaded.

The package allocates a fixed number of these protected objects. The
function Load blocks until it can acquire one of these objects. Using
smart pointers, a Future is automatically released (so that it can be
re-acquired for loading another resource) when it is no longer
referenced.

Signed-off-by: onox <denkpadje@gmail.com>