orka: Refactor package Smart_Pointers to actually store a pointer
Previously an Abstract_Pointer in Smart_Pointers would store a definite
value instead of a pointer.

This meant that a reference returned via Get would have a discriminant
pointing to the value. If this value happens to be a raw pointer, then
using Get would mean the reference points to a raw pointer, which then
points to some object. This required writing Get.all instead of just
Get.

By actually storing a pointer instead of a definite value, the reference
can directly point to the object.

Signed-off-by: onox <denkpadje@gmail.com>