examples: Optimize compute example by unrolling some iterations
The optimization assumes that the size of a sub-group (a number of
invocations that execute in lockstep) is 32 or higher.

The actual size of the sub-group is written to the SSBO at the end of
the shader using the ARB_shader_ballot extension. If the extension is
not available, it can be safely commented.

Signed-off-by: onox <denkpadje@gmail.com>