orka: If future has failed, don't execute remaining jobs in job graph
If execution of a job fails, any jobs (and their dependents) appended by
the job to the Jobs vector will be enqueued and dequeued by executors.
However, executors will not execute these jobs since the future has
status failed. This is necessary to make sure that all jobs in the graph
are freed.

Signed-off-by: onox <denkpadje@gmail.com>