orka: Add generic parameter Index_Type to package Bounded_Vectors
An explicit type for the index is needed in order to have arrays of
resources and handles in package Orka.Frame_Graphs with incompatible
index types.

This avoids bugs like mixing offsets in arrays of handles with the
handles themselves, which would result in being unable to construct
valid graphs or incorrect culling of graphs.

Signed-off-by: onox <denkpadje@gmail.com>