orka: Add implicit resource when adding a resource as an attachment
Resources which are read (input) or written (output) as a framebuffer
attachment will have a corresponding output or input in the graph. For
example, a graph with one pass writing to a color attachment:

   Pass 1 -> Resource 1

Becomes:

   Resource 1 -> Pass 1 -> Resource 1

This fixes the logic which determines which color attachments needs to
be cleared. It also simplifies determining whether writes to a a depth
or stencil attachment needs to be prevented by checking whether the
output attachment is implicit or not.

Signed-off-by: onox <denkpadje@gmail.com>