orka: Create framebuffers and attachment textures for frame graph
Add a present pass in function Cull and move construction of
framebuffers and textures that are going to be attached to them to
procedure Initialize.

Disable depth or stencil writes if the output resource is implicit.

There are three ways to present a resource to the screen:

 * Using the default framebuffer during the last render pass if it has
   one color attachment (actual present pass will do nothing)

 * Blit the resource to the default framebuffer if the last pass has
   multiple color attachment, if the presented resource is a depth
   or stencil attachment, or if the resource is not a Texture_2D

 * Render the resource using the default framebuffer if it was not
   written by the last render pass as a framebuffer attachment

Currently only the first method is implemented. Input resources that are
not framebuffer attachments should probably be created where they can be
binded to the shader program's uniforms of a render pass.

Function Get_Texture may get remove in a future commit when a transient
resource system gets implemented.

Signed-off-by: onox <denkpadje@gmail.com>