orka: Workaround GLFW bug giving weird mouse position after locking
If the user unlocks the pointer and stops moving the mouse, and then
locks the pointer *and then* moves the mouse, then the first new mouse
position in the Mouse_Position_Changed callback has a weird value,
causing large deltas. This results in the camera instantly rotating back
to an old position.

This behavior does not occur if the user moves the mouse first and,
while moving the mouse, locks the pointer.

Could not reproduce this bug with SDL.

Signed-off-by: onox <denkpadje@gmail.com>