orka: Add packages for joysticks and gamepads
A Joystick object can be created with:

   JS : Orka.Inputs.Joysticks.Joystick_Input_Access :=
     Orka.Inputs.GLFW.Create_Joystick_Input (Index);

where Index is some Glfw.Input.Joysticks.Joystick_Index.

To update the state of the joystick each frame, call Update_State:

   JS.Update_State (null);

Or provide an access-to-procedure value to process the axes:

   declare
      package Filtering renames Orka.Inputs.Joysticks.Filtering;

      procedure Process_Axis
        (Value : in out Orka.Inputs.Joysticks.Axis_Position;
         Index :        Positive) is
      begin
         Value := Filtering.Dead_Zone (Value, 0.05);
      end Process_Axis;
   begin
      JS.Update_State (Process_Axis'Access);
   end;

This procedure can be used to handle the dead zone of an axis, to invert
it, and/or to smooth its signal with a low-pass filter.

After the state has been updated, it can be retrieved with the function
Current_State. The previous state can be retrieved with Last_State.

Signed-off-by: onox <denkpadje@gmail.com>