orka: Move checking query result of timer to procedure Start
Issue 1:

Moving the checking of the availability of the result of the query to
the procedure Start, instead of in the function GPU_Duration, will make
sure that, on most systems, a new result is available every frame
instead of every two frames.

The timer is started with procedure Start and then later stopped with
procedure Stop. Shortly afterwards, function GPU_Duration may be called
to update any UI. At this time, the result is usually not available yet.
Therefore, the next call to Start and Stop (in frame i + 1) will not do
anything and we have to wait until GPU_Duration is called again to get
the result of frame i. In the next frame, i + 2, we can start a new
query.

However, especially if vertical sync is enabled, enough time may have
passed for the result of frame i to become available when procedure
Start is called in frame i + 1. After obtaining the result, we can
immediately start a new query.

Issue 2:

The timestamp returned by OpenGL may overflow much sooner than its type
suggests. Check that the stop time is greater than the start time and
skip updating the time otherwise.

Signed-off-by: onox <denkpadje@gmail.com>