orka: Replace procedures Set_Texture and Set_Image with null procedures
Procedures Set_Texture and Set_Image suggest that a texture can be
binded to a uniform/image sampler and then be forgotten.
However, textures are actually binded to a binding point at some index.
This means it would be more appropriate to require explicit bindings in
the shaders and some procedure to bind textures to these binding points.

The mentioned procedures have been replaced with two null procedures
that use a pre condition to verify that the kind (1D, 2D, etc.) and
format of the texture and sampler match.

Signed-off-by: onox <denkpadje@gmail.com>