gl: Dereference the actual GL pointer instead of looking it up in a map
Generic subprograms in GL.API are now generic packages containing a
variable Ref which points to a helper subprogram. This subprogram loads
the actual GL function pointer and stores it in Ref (replacing the old
pointer which pointed to the helper subprogram).

The second time that the variable Ref is dereferenced, it directly calls
the actual GL function. This completely avoids having to look up GL
function pointers in a hash map.

Signed-off-by: onox <denkpadje@gmail.com>