orka: Remove type Vertex_Format
Vertices can be pulled programmatically in vertex shaders. This avoids
having to track and pass around Vertex_Format objects.

For example, if the vertex data consists of:

 * vec3 position (as 3x 16-bit Half_Type)
 * vec3 normal   (as 3x 16-bit Half_Type)
 * vec2 uv       (as 2x 16-bit Half_Type)

Then the buffer containing the data can be binded as an SSBO:

  layout(std430, binding = 0) readonly restrict buffer vertexBuffer {
      uvec4 in_vertices[];
  };

The data can then be pulled and unpacked with 4x unpackHalf2x16():

  const uvec4 data = in_vertices[gl_VertexID];

  const vec2 tmp = unpackHalf2x16(data.y);
  const vec3 in_Position = vec3(unpackHalf2x16(data.x), tmp.x);
  const vec3 in_Normal = vec3(tmp.y, unpackHalf2x16(data.z));
  const vec2 in_UV = unpackHalf2x16(data.w);

Signed-off-by: onox <denkpadje@gmail.com>
Fixes: #60