orka: Add function Create_Context with a native Wayland display pointer
Type Wayland_EGL_Context requires a pointer to a native Wayland display,
otherwise EGL will fail to attach a Wayland buffer to a Wayland surface
when swapping buffers.

Extension EXT_platform_wayland says a display can be constructed using
EGL_DEFAULT_DISPLAY, but apparently this is not the case with Mesa.

Signed-off-by: onox <denkpadje@gmail.com>