egl: Do not uncurrent context in Finalize if not current on current task
If some task tries to finalize an EGL context, it should not attempt to
make the context not current on the current task if the context is not
current on that task.

If the context was made current in some task, then this task may want
to call Make_Not_Current before this task terminates.

Signed-off-by: onox <denkpadje@gmail.com>