awt: Add support for gamepads
Package AWT.Inputs.Gamepads provides several subprograms to set gamepad
mappings from [1], initialize and retrieve the currently connected
gamepads, and to listen for (dis)connection events.

Each gamepad has a name, unique serial number, a GUID, and a set of
buttons, axes, and triggers. The state of the battery and LED, if
present, can be retrieved. The color of the LED can be set as well.

Some subprograms exist to play and cancel rumble and periodic
force-feedback effects.

[1] https://github.com/gabomdq/SDL_GameControllerDB

Signed-off-by: onox <denkpadje@gmail.com>