awt: Estimate orientation of gamepad using motion sensor
State estimation using the square-root CDKF Kalman filter. The measured
angular velocity from the gyro is used as the input in the state
transition function to compute the estimated orientation.

The estimated orientation is used to compute how the gravity vector
would be measured, and thus only the acceleration is used as a
measurement.

The gyro bias is automatically calibrated using a number of measurements
while the gamepad is not moved or rotated.

Signed-off-by: onox <denkpadje@gmail.com>
Fixes: #80