numerics: Implement functions Quantile and Median
The function Quantile uses the quickselect selection algorithm using the
median-of-3 as the pivot strategy. This can still give a worst-case
performance of O(n^2) for some inputs.

Further improvements can be made by using Tukey's ninther and/or by
falling back to the modified median-of-medians algorithm [1] when the
quickselect algorithm goes quadratic.

[1] https://arxiv.org/abs/1606.00484

Signed-off-by: onox <denkpadje@gmail.com>