orka: Fix compilation of package Rendering.Debug.Spheres
Most of the getters to query the current rendering state of GL have been
removed since querying can be slow and objects that implement the
interface Context already keep track of the current state if Context's
procedure Update_State is called.

The packages Rendering.Debug.* need to be rewritten to return a frame
graph instead of directly modifying GL state and executing draw calls.

Signed-off-by: onox <denkpadje@gmail.com>