orka: Add some commented subprograms to Orka.Frame_Graphs
To make a frame graph really usable, it should be able to return a list
of exported and imported resources (the outputs and inputs). The object
that is returned should somehow contain a reference to its frame graph.

An exported resource can then be given as a parameter to procedure
Add_Input and an imported resource as a parameter to procedure
Add_Output of a second frame graph.

This would allow packages to return a frame graph that applies certain
operations to its inputs. The frame graph can then be inserted into the
main (bigger) frame graph.

Signed-off-by: onox <denkpadje@gmail.com>